requirements.txt
Flask
gunicorn
numpy
scikit-learn
Werkzeug
Jinja2
itsdangerous
click
scipy
joblib
threadpoolctl
Flask-CORS
requests
markdown
google-generativeai
pandas
python-dotenv
markupsafe